apps/api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.10.5
pydantic-settings==2.8.1
pytest==8.3.3
pytest-asyncio==0.23.8
httpx==0.27.0
PyJWT==2.10.1
cryptography==44.0.0